[de] minor bugfix: correctly handle capitalization of added words
Words that were added to 'de/hunspell/spelling.txt' are case-sensitive.
For instance, if spelling.txt contains 'holomorph/A', the word 
'Holomorphe' will be marked as incorrect. This is fine, unĺess it is at
the sentence start (e.g. 'Holomorphe Funktionen sind doof.'). 
In this case, also the capitalized word has to be accepted.